Implement _even more modern_ VM structure.
