+++ app/src/main/java/com/kickstarter/ui/activities/ProjectActivity.java
@@ -24,7 +24,6 @@ import com.kickstarter.ui.IntentKey;
 import com.kickstarter.ui.adapters.ProjectAdapter;
 import com.kickstarter.ui.data.LoginReason;
 import com.kickstarter.ui.views.IconButton;
-import com.kickstarter.viewmodels.ProjectViewModel;
 
 import javax.inject.Inject;
 
@@ -36,8 +35,8 @@ import butterknife.ButterKnife;
 import butterknife.OnClick;
 import rx.android.schedulers.AndroidSchedulers;
 
-@RequiresActivityViewModel(ProjectViewModel.class)
-public final class ProjectActivity extends BaseActivity<ProjectViewModel> {
+@RequiresActivityViewModel(com.kickstarter.viewmodels.Project.ViewModel.class)
+public final class ProjectActivity extends BaseActivity<com.kickstarter.viewmodels.Project.ViewModel> {
   private ProjectAdapter adapter;
 
   protected @Bind(R.id.project_recycler_view) RecyclerView projectRecyclerView;
