--- app/src/main/java/com/kickstarter/viewmodels/inputs/ProjectViewModelInputs.java
@@ -1,18 +1,0 @@
-package com.kickstarter.viewmodels.inputs;
-
-public interface ProjectViewModelInputs {
-  /**
-   * Call when the star button has been clicked.
-   */
-  void starClicked();
-
-  void backProjectClicked();
-  void shareClicked();
-  void blurbClicked();
-  void commentsClicked();
-  void creatorNameClicked();
-  void managePledgeClicked();
-  void updatesClicked();
-  void playVideoClicked();
-  void viewPledgeClicked();
-}
